Organize parser
