+++ quickapi/http/request/__init__.py
@@ -2,7 +2,7 @@ import attrs
 
 from quickapi.http.body import Body
 from quickapi.http.request.method import Method
-from quickapi.http.request.parser import from_connection
+from quickapi.http.request.parse import from_connection
 from quickapi.http.request.target import Target
 from quickapi.http.version import Version
